Prevent hiding from your past
Previously, ancestors spawned wandering in some circumstances, and
stealthy Heplites could sometimes have their ancestor wander off.

This happened because we set the monster's foe to a previous target of
shout-attack, but then it is out of LoS, so it the foe gets cleared
altogether rather than set to the player.

This also affected at least Mak demons, though it is much less likely
to come up - the flow where it usually happens for ancestors is when
you bail from a fight when your ancestor dies.

It did not affect regular summons, possibly only because spellcasting
triggers an extra alert to every monster in your LoS.